Reorganize cypher builder tests
